# openapi.json
{"openapi":"3.1.0","info":{"title":"Bikube - API for kommuniksjon med tekstkataloger","description":"REST-API for kommunikasjon mot kataloger for tekstmateriale. Brukes av produksjonsløypene for å lettere kunne hente ut katalog-data enn ved direkteintegrering. ","version":"0.0.1-SNAPSHOT"},"servers":[{"url":"","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Newspaper titles","description":"Endpoints related to newspaper titles."},{"name":"Newspaper items","description":"Endpoints related to newspaper items."},{"name":"MarcXchange HTML","description":"Get bibliographic records as MarcXchange encoded MARC-21 in HTML"},{"name":"MarcXchange","description":"Get bibliographic records as MarcXchange encoded MARC-21"},{"name":"Catalogue objects","description":"Endpoints related to catalog data for all text material"}],"paths":{"/api/newspapers/items":{"put":{"tags":["Newspaper items"],"summary":"Update a single physical newspaper by manifestation ID.","description":"Update newspaper. Notes and number is optional (not ID and username), and fields not provided will not be updated. All possible fields are only present on the manifestation in the current catalog, and will be updated there.Date cannot be changed. In case of date change, delete the physical item and create a new one.","operationId":"updateItem","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemUpdateDto"}}},"required":true},"responses":{"204":{"description":"Newspaper item updated"},"400":{"description":"Bad request"},"404":{"description":"Not found"},"500":{"description":"Server error"}}},"post":{"tags":["Newspaper items"],"summary":"Create a single newspaper item","operationId":"createItem","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemInputDto"}}},"required":true},"responses":{"201":{"description":"Newspaper created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Item"}}}},"400":{"description":"Bad request"},"409":{"description":"Conflict"},"500":{"description":"Server error"}}}},"/api/hugin/titles/contact":{"put":{"tags":["hugin-newspaper-controller"],"operationId":"upsertContactInformation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactUpdateDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HuginTitle"}}}}}}},"/api/newspapers/titles/":{"post":{"tags":["Newspaper titles"],"summary":"Create a newspaper title","operationId":"createTitle","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TitleInputDto"}}},"required":true},"responses":{"201":{"description":"Title created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Title"}}}},"400":{"description":"Bad request"},"500":{"description":"Server error"}}}},"/api/newspapers/items/missing":{"post":{"tags":["Newspaper items"],"summary":"Create a 'missing item', in other words create a manifestation without item","operationId":"createManifestationOnly","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MissingPeriodicalItemDto"}}},"required":true},"responses":{"201":{"description":"Manifestation/missing item created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Item"}}}},"400":{"description":"Bad request"},"500":{"description":"Server error"}}}},"/api/hugin/newspapers/batch":{"post":{"tags":["hugin-newspaper-controller"],"operationId":"upsertNewspapers","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NewspaperUpsertDto"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Newspaper"}}}}}}}},"/api/hugin/box":{"post":{"tags":["hugin-newspaper-controller"],"operationId":"createBox","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBoxDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Box"}}}}}}},"/":{"get":{"tags":["swagger-redirect-controller"],"operationId":"root","responses":{"200":{"description":"OK"}}}},"/api/title/search":{"get":{"tags":["Catalogue objects"],"summary":"Search catalogue titles","description":"Searches catalogue titles by name. Supports wildcard '*' search.","operationId":"searchTitle","parameters":[{"name":"searchTerm","in":"query","description":"Search term","required":true,"schema":{"type":"string"}},{"name":"materialType","in":"query","description":"Material type of the titles to search for","required":true,"schema":{"type":"string","enum":["NEWSPAPER","MANUSCRIPT","PERIODICAL","MONOGRAPH"]}},{"name":"date","in":"query","description":"Date in ISO-8601 format (YYYY-MM-DD). If provided, only titles with start and end date that includes this date will be returned.","required":false,"schema":{"type":"string","format":"date"}},{"name":"selectBestMatch","in":"query","description":"Whether or not to select only the best match for search term","required":false,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"oneOf":[{"$ref":"#/components/schemas/Item"},{"$ref":"#/components/schemas/Language"},{"$ref":"#/components/schemas/Publisher"},{"$ref":"#/components/schemas/PublisherPlace"},{"$ref":"#/components/schemas/Title"}]}}}}},"400":{"description":"Bad request"},"500":{"description":"Server error"}}}},"/api/title/link":{"get":{"tags":["Catalogue objects"],"summary":"Get a direct link to title in catalogue","operationId":"getCatalogueLink","parameters":[{"name":"catalogueId","in":"query","required":true,"schema":{"type":"string"}},{"name":"materialType","in":"query","required":true,"schema":{"type":"string","enum":["NEWSPAPER","MANUSCRIPT","PERIODICAL","MONOGRAPH"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"url"}}}},"400":{"description":"Bad request"}}}},"/api/title":{"get":{"tags":["Catalogue objects"],"summary":"Get single title from catalogue","operationId":"getSingleTitle","parameters":[{"name":"catalogueId","in":"query","required":true,"schema":{"type":"string"}},{"name":"materialType","in":"query","required":true,"schema":{"type":"string","enum":["NEWSPAPER","MANUSCRIPT","PERIODICAL","MONOGRAPH"]}},{"name":"useSeries","in":"query","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Title"}}}},"400":{"description":"Bad request"},"500":{"description":"Server error"}}}},"/api/item/search":{"get":{"tags":["Catalogue objects"],"summary":"Search catalogue items","description":"Search catalogue items by title id, material type, date and digital/physical.","operationId":"searchItem","parameters":[{"name":"titleCatalogueId","in":"query","description":"Catalogue ID of the title to search items for.","required":true,"schema":{"type":"string"}},{"name":"materialType","in":"query","description":"Material type of the items to search for.","required":true,"schema":{"type":"string","enum":["NEWSPAPER","MANUSCRIPT","PERIODICAL","MONOGRAPH"]}},{"name":"date","in":"query","description":"Date in ISO 8601 format (YYYY-MM-DD).","required":true,"schema":{"type":"string","pattern":"^(17|18|19|20)\\d{2}(-)?(0[1-9]|1[0-2])(-)?(0[1-9]|[12][0-9]|3[01])$"}},{"name":"isDigital","in":"query","description":"If 'true' returns only digital items. If 'false', returns only physical items. Default is all.","required":false,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"oneOf":[{"$ref":"#/components/schemas/Item"},{"$ref":"#/components/schemas/Language"},{"$ref":"#/components/schemas/Publisher"},{"$ref":"#/components/schemas/PublisherPlace"},{"$ref":"#/components/schemas/Title"}]}}}}},"400":{"description":"Bad request"},"500":{"description":"Server error"}}}},"/api/item":{"get":{"tags":["Catalogue objects"],"summary":"Get single item from catalogue","operationId":"getSingleItem","parameters":[{"name":"catalogueId","in":"query","required":true,"schema":{"type":"string"}},{"name":"materialType","in":"query","required":true,"schema":{"type":"string","enum":["NEWSPAPER","MANUSCRIPT","PERIODICAL","MONOGRAPH"]}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Item"}}}},"500":{"description":"Server error"}}}},"/api/hugin/titles/{titleId}":{"get":{"tags":["hugin-newspaper-controller"],"operationId":"getTitle","parameters":[{"name":"titleId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HuginTitle"}}}}}}},"/api/collections/raw":{"get":{"tags":["collections-raw-controller"],"operationId":"getRawCollections","parameters":[{"name":"params","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/auth/me":{"get":{"tags":["auth-controller"],"operationId":"getUserInfo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/alma/search":{"get":{"tags":["MarcXchange"],"summary":"Find records via free text search","operationId":"searchAlma","parameters":[{"name":"title","in":"query","required":true,"schema":{"type":"string"}},{"name":"author","in":"query","required":false,"schema":{"type":"string"}},{"name":"year","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/xml":{"schema":{"type":"string","format":"byte"}}}},"400":{"description":"Invalid query"},"500":{"description":"Server error"}}}},"/api/alma/mms/{mms}":{"get":{"tags":["MarcXchange"],"summary":"Get bibliographic record by title MMS-ID.","operationId":"getAlmaItemByMMS","parameters":[{"name":"mms","in":"path","description":"Title MMS-ID","required":true,"schema":{"type":"string","pattern":"[0-9]{8,19}"}},{"name":"prolog","in":"query","description":"Include XML declaration/prolog in XML output.","required":false,"schema":{"type":"boolean","default":true}}],"responses":{"200":{"description":"OK","content":{"application/xml":{"schema":{"type":"string","format":"byte"}}}},"400":{"description":"Invalid MMS","content":{"application/xml":{"schema":{"type":"string","format":"byte"}}}},"404":{"description":"MMS not found","content":{"application/xml":{"schema":{"type":"string","format":"byte"}}}},"500":{"description":"Server error","content":{"application/xml":{"schema":{"type":"string","format":"byte"}}}}}}},"/api/alma/issn/{issn}":{"get":{"tags":["MarcXchange"],"summary":"Get bibliographic records by ISSN.","operationId":"getMarcRecordsByISSN","parameters":[{"name":"issn","in":"path","description":"ISSN","required":true,"schema":{"type":"string","pattern":"[0-9]{4}-?[0-9]{3}[0-9Xx]"}}],"responses":{"200":{"description":"OK","content":{"application/xml":{"schema":{"type":"string","format":"byte"}}}},"400":{"description":"Invalid ISSN","content":{"application/xml":{"schema":{"type":"string","format":"byte"}}}},"404":{"description":"ISSN not found","content":{"application/xml":{"schema":{"type":"string","format":"byte"}}}},"500":{"description":"Server error","content":{"application/xml":{"schema":{"type":"string","format":"byte"}}}}}}},"/api/alma/ismn/{ismn}":{"get":{"tags":["MarcXchange"],"summary":"Get bibliographic records by ISMN.","operationId":"getMarcRecordsByISMN","parameters":[{"name":"ismn","in":"path","description":"ISMN","required":true,"schema":{"type":"string","pattern":"([0-9]-?){9}[0-9](-?(([0-9]-?){2}[0-9]))?|M-?[0-9]{4}-?[0-9]{4}-?[0-9]"}}],"responses":{"200":{"description":"OK","content":{"application/xml":{"schema":{"type":"string","format":"byte"}}}},"400":{"description":"Invalid ISMN"},"404":{"description":"ISMN not found"},"500":{"description":"Server error"}}}},"/api/alma/isbn/{isbn}":{"get":{"tags":["MarcXchange"],"summary":"Get bibliographic records by ISBN.","operationId":"getMarcRecordsByISBN","parameters":[{"name":"isbn","in":"path","description":"ISBN","required":true,"schema":{"type":"string","pattern":"([0-9]-?){9}[0-9](-?(([0-9]-?){2}[0-9]))?"}}],"responses":{"200":{"description":"OK","content":{"application/xml":{"schema":{"type":"string","format":"byte"}}}},"400":{"description":"Invalid ISBN"},"404":{"description":"ISBN not found"},"500":{"description":"Server error"}}}},"/api/alma/html/barcode/{barcode}":{"get":{"tags":["MarcXchange HTML"],"summary":"Get bibliographic record by barcode","description":"Returns a bibliographic record in HTML format for the given barcode","operationId":"getHTMLByBarcode","parameters":[{"name":"barcode","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/alma/barcode/{barcode}":{"get":{"tags":["MarcXchange"],"summary":"Get bibliographic record by item barcode.","operationId":"getAlmaItemByBarcode","parameters":[{"name":"barcode","in":"path","description":"Item barcode","required":true,"schema":{"type":"string","pattern":"[a-zA-Z0-9]{4,13}"}},{"name":"prolog","in":"query","description":"Include XML declaration/prolog in XML output.","required":false,"schema":{"type":"boolean","default":true}}],"responses":{"200":{"description":"OK","content":{"application/xml":{"schema":{"type":"string","format":"byte"}}}},"400":{"description":"Invalid barcode","content":{"application/xml":{"schema":{"type":"string","format":"byte"}}}},"404":{"description":"Barcode not found","content":{"application/xml":{"schema":{"type":"string","format":"byte"}}}},"500":{"description":"Server error","content":{"application/xml":{"schema":{"type":"string","format":"byte"}}}}}}},"/api/newspapers/items/physical/{id}":{"delete":{"tags":["Newspaper items"],"summary":"Delete a single physical newspaper by ID of manifestation.","description":"Delete a single physical newspaper by the ID of its manifestation.Default, the manifestation will be deleted if there are no other items attached, or if there are no items on the manifestation. Can be turned off by setting deleteManifestation=false. When set to true, will still delete item (not manifestation) and return OK if there are more items on the manifestation.","operationId":"deleteItem","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"deleteManifestation","in":"query","description":"Also delete manifestation if no other items are attached to it.","required":false,"schema":{"type":"boolean","default":true}}],"responses":{"204":{"description":"Newspaper item deleted"},"400":{"description":"Bad request"},"404":{"description":"Not found"},"500":{"description":"Server error"}}}},"/api/hugin/newspapers/{manifestationId}":{"delete":{"tags":["hugin-newspaper-controller"],"operationId":"deleteNewspaper","parameters":[{"name":"manifestationId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"ItemUpdateDto":{"type":"object","properties":{"manifestationId":{"type":"string"},"username":{"type":"string"},"notes":{"type":"string"},"number":{"type":"string"}},"required":["manifestationId","username"]},"ContactInfoDto":{"type":"object","properties":{"contactType":{"type":"string","enum":["phone","email"]},"contactValue":{"type":"string"}},"required":["contactType","contactValue"]},"ContactUpdateDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"vendor":{"type":"string"},"contactName":{"type":"string"},"shelf":{"type":"string"},"notes":{"type":"string"},"contactInfos":{"type":"array","items":{"$ref":"#/components/schemas/ContactInfoDto"}},"releasePattern":{"type":"array","items":{"type":"integer","format":"int32"}}},"required":["id"]},"Box":{"type":"object","properties":{"id":{"type":"string"},"dateFrom":{"type":"string","format":"date"},"active":{"type":"boolean"},"title":{"$ref":"#/components/schemas/HuginTitle"},"newspapers":{"type":"array","items":{"$ref":"#/components/schemas/Newspaper"}}},"required":["active","id","newspapers"]},"ContactInfo":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"title":{"$ref":"#/components/schemas/HuginTitle"},"contactType":{"type":"string","enum":["phone","email"]},"contactValue":{"type":"string"}}},"HuginTitle":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"contactName":{"type":"string"},"vendor":{"type":"string"},"releasePattern":{"type":"array","items":{"type":"integer","format":"int32"}},"shelf":{"type":"string"},"notes":{"type":"string"},"boxes":{"type":"array","items":{"$ref":"#/components/schemas/Box"}},"contactInfos":{"type":"array","items":{"$ref":"#/components/schemas/ContactInfo"}}},"required":["boxes","contactInfos","id"]},"Newspaper":{"type":"object","properties":{"edition":{"type":"string"},"date":{"type":"string","format":"date"},"received":{"type":"boolean"},"username":{"type":"string"},"notes":{"type":"string"},"box":{"$ref":"#/components/schemas/Box"},"catalogId":{"type":"string"}},"required":["catalogId"]},"TitleInputDto":{"type":"object","properties":{"name":{"type":"string"},"username":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"publisher":{"type":"string"},"publisherPlace":{"type":"string"},"language":{"type":"string"}},"required":["name","username"]},"RelatedTitle":{"type":"object","properties":{"catalogueId":{"type":"string"},"title":{"type":"string"},"association":{"type":"string"},"recordType":{"type":"string"}}},"Title":{"allOf":[{"$ref":"#/components/schemas/CatalogueRecord"},{"type":"object","properties":{"catalogueId":{"type":"string"},"name":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"publisher":{"type":"string"},"publisherPlace":{"type":"string"},"language":{"type":"string"},"materialType":{"type":"string"},"relatedTitles":{"type":"array","items":{"$ref":"#/components/schemas/RelatedTitle"}}}}],"required":["catalogueId"]},"MissingPeriodicalItemDto":{"type":"object","properties":{"date":{"type":"string","format":"date"},"titleCatalogueId":{"type":"string"},"username":{"type":"string"},"notes":{"type":"string"},"volume":{"type":"string"},"number":{"type":"string"},"version":{"type":"string"}},"required":["date","titleCatalogueId","username"]},"Item":{"allOf":[{"$ref":"#/components/schemas/CatalogueRecord"},{"type":"object","properties":{"catalogueId":{"type":"string"},"name":{"type":"string"},"date":{"type":"string","format":"date"},"materialType":{"type":"string"},"titleCatalogueId":{"type":"string"},"titleName":{"type":"string"},"digital":{"type":"boolean"},"urn":{"type":"string"},"location":{"type":"string"},"parentCatalogueId":{"type":"string"}}}],"required":["catalogueId"]},"ItemInputDto":{"type":"object","properties":{"date":{"type":"string","format":"date"},"titleCatalogueId":{"type":"string"},"username":{"type":"string"},"digital":{"type":"boolean"},"urn":{"type":"string"},"containerId":{"type":"string"},"notes":{"type":"string"},"number":{"type":"string"},"volume":{"type":"string"},"version":{"type":"string"},"itemStatus":{"type":"string"}},"required":["date","titleCatalogueId","username"]},"NewspaperUpsertDto":{"type":"object","properties":{"titleId":{"type":"integer","format":"int32"},"boxId":{"type":"string"},"date":{"type":"string","format":"date"},"edition":{"type":"string"},"received":{"type":"boolean"},"notes":{"type":"string"},"catalogId":{"type":"string"}},"required":["boxId","date","received"]},"CreateBoxDto":{"type":"object","properties":{"titleId":{"type":"integer","format":"int32"},"id":{"type":"string"},"dateFrom":{"type":"string","format":"date"}},"required":["dateFrom","id","titleId"]},"CatalogueRecord":{},"Language":{"allOf":[{"$ref":"#/components/schemas/CatalogueRecord"},{"type":"object","properties":{"name":{"type":"string"},"catalogueId":{"type":"string"}}}]},"Publisher":{"allOf":[{"$ref":"#/components/schemas/CatalogueRecord"},{"type":"object","properties":{"name":{"type":"string"},"catalogueId":{"type":"string"}}}]},"PublisherPlace":{"allOf":[{"$ref":"#/components/schemas/CatalogueRecord"},{"type":"object","properties":{"name":{"type":"string"},"catalogueId":{"type":"string"}}}]},"User":{"type":"object","properties":{"username":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}}},"required":["roles","username"]}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}